+++ lib/fl_webview.dart
@@ -1,6 +1,5 @@
 library fl_webview;
 
 export 'src/adapt_web_view.dart';
-export 'src/method_channel.dart';
 export 'src/platform_interface.dart';
 export 'src/web_view.dart';